+++ crates/vertigo/src/driver_module/src_js/exec_command/command/dom/features.ts
@@ -1,0 +1,6 @@
+
+export const getDisableHydration = (): boolean => {
+    const metadataDiv = document.getElementById('v-metadata');
+    const value = metadataDiv?.getAttribute('data-env-disable-hydration');
+    return value === 'true';
+};