+++ CHANGES.md
@@ -9,6 +9,7 @@
 * `AutoJsJson`: Added `rename_all` container attribute, and `rename` field attribute [#406]
 * `AutoJsJson`: Support for `JsJson` type for dynamic schema [#393]
 * `SsrFetchCache` - Cache passed in HTML so the browser doesn't need to refetch the data already used during SSR [#413] [#414]
+* Hydration (no unnecessary DOM nodes re-creation after page load)
 
 ### Fixed